Remove custom cover image field
We have a button for this now, and it's not worth the effort to do additional patch checking to see if the user has hosted something somewhere else. If a custom thumbnail wanted to be used, the user can upload it, and then unapprove the image in the gallery, thus allowing it to be used as a thumbnail.

closes #7